fix: check if pyproject has poetry earlier
During the evaluation in the package phase, we determine whether a
`requirements.txt` file exists, or whether we need to generate one.

Since the `pyproject.toml` file is used by poetry, but only if a stanza
is contained inside the file, use the function `isPoetryProject()` along
with the configuration value, thereby reducing the need for a project to
have to declare a configuration override.

Refs #324
Refs #344
Fixes #400

Signed-off-by: Mike Fiedler <miketheman@gmail.com>